CHG: ZarrDragAndDropOpenSettingsUI move to 'plugins' package:
- it is now together with all other IJ GUI plugins
- was also renamed to DragAndDropBehaviorSettings,
  and so was renamed also its Test class
- references to the new name were updated in the Test class
